requirements.txt
jax[tpu]
absl-py
flax
jaxlib
numpy
jsonlines
ray[default]
# For profiling
tensorboard-plugin-profile
# Install tensorflow to avoid spurring "Can't import tensorflow.python.profiler.trace" error from jax.profiler.
tensorflow
google-cloud-storage
pandas
openpyxl
